new fmt rules

+++ test/foundry/22_HashVerificationTest.t.sol
@@ -49,7 +49,7 @@ contract HashVerificationTest is TestUtils {
     string public constant TEST_SOLVER_ADDRESS = "0x0999CB4Ead0E01C861c2Bfe4B31130185a3adfA5";
     string public constant TEST_SOLVER_PRIVATE_KEY = "c0b4a772933191e90eadddbb6cade5f6c47abff65c5c7c92869d9444ef18750a";
     uint256 private constant SOLVER_PRIVATE_KEY = 0xc0b4a772933191e90eadddbb6cade5f6c47abff65c5c7c92869d9444ef18750a;
-    
+
     address public testSigner;
     address public solverAddress;
 
@@ -89,26 +89,26 @@ contract HashVerificationTest is TestUtils {
         // Calculate signing hash using the Arbitrum contract address
         // This is the actual hash that would be signed in production
         bytes32 signingHash = calculateSigningHashWithAddress(order, ARBITRUM_CONTRACT_ADDRESS);
-        
+
         // Generate signature using offerer's private key
         (uint8 rawV, bytes32 r, bytes32 s) = vm.sign(OFFERER_PRIVATE_KEY, signingHash);
-        
+
         // Verify signature recovery using ecrecover
         address recovered = ecrecover(signingHash, rawV, r, s);
         assertEq(recovered, testSigner, "Signature recovery failed");
-        
+
         // Create a properly formatted signature - for Solady ECDSA, v must be 27/28
         bytes memory signature = new bytes(65);
-        
+
         // Copy r, s
-        for (uint i = 0; i < 32; i++) {
+        for (uint256 i = 0; i < 32; i++) {
             signature[i] = bytes32ToBytes(r)[i];
             signature[32 + i] = bytes32ToBytes(s)[i];
         }
-        
+
         // Set v (must be 27 or 28 for Solady ECDSA)
         signature[64] = bytes1(rawV);
-        
+
         // Print relevant information
         console.log("\n==== HASH VERIFICATION TEST ====");
         console.log("Expected Signer:        %s", testSigner);
@@ -116,7 +116,7 @@ contract HashVerificationTest is TestUtils {
         console.log("Solver Address:         %s", solverAddress);
         console.log("\n---- ORDER DETAILS ----");
         console.log("Offerer:                %s", order.offerer);
-        console.log("Recipient:              %s", order.recipient); 
+        console.log("Recipient:              %s", order.recipient);
         console.log("Input Token:            %s", order.inputToken);
         console.log("Output Token:           %s", order.outputToken);
         console.log("Input Amount:           %d", uint256(order.inputAmount));
@@ -125,25 +125,21 @@ contract HashVerificationTest is TestUtils {
         console.log("End Time:               %d", order.endTime);
         console.log("Source EID:             %d", order.srcEid);
         console.log("Destination EID:        %d", order.dstEid);
-        
+
         // ==================== DEPOSIT VERIFICATION STEP ====================
         // We'll deploy a local test contract at the same address as the Arbitrum contract
         // to verify that our signature works correctly for deposits
         console.log("\n---- VERIFYING DEPOSIT WITH TEST CONTRACT ----");
-        
+
         // Get the code for a new contract with Arbitrum's EID
         // Note: This test manually deploys impl+proxy to use vm.etch for address manipulation
         address layerZeroEndpoint = address(endpoints[1]); // Use test endpoint from TestUtils
         Aori implementation = new Aori(layerZeroEndpoint, ARBITRUM_EID);
         ERC1967Proxy proxy = new ERC1967Proxy(
             address(implementation),
-            abi.encodeCall(implementation.initialize, (
-                address(this),
-                MAX_FILLS_PER_SETTLE,
-                new address[](0),
-                new address[](0),
-                new uint32[](0)
-            ))
+            abi.encodeCall(
+                implementation.initialize, (address(this), MAX_FILLS_PER_SETTLE, new address[](0), new address[](0), new uint32[](0))
+            )
         );
 
         // Deploy the test contract at the exact Arbitrum address using vm.etch
@@ -156,14 +152,8 @@ contract HashVerificationTest is TestUtils {
         vm.store(ARBITRUM_CONTRACT_ADDRESS, implSlot, bytes32(uint256(uint160(address(implementation)))));
 
         // Initialize the etched contract
-        Aori(ARBITRUM_CONTRACT_ADDRESS).initialize(
-            address(this),
-            MAX_FILLS_PER_SETTLE,
-            new address[](0),
-            new address[](0),
-            new uint32[](0)
-        );
-        
+        Aori(ARBITRUM_CONTRACT_ADDRESS).initialize(address(this), MAX_FILLS_PER_SETTLE, new address[](0), new address[](0), new uint32[](0));
+
         // ADD THIS CODE HERE - before any deposit operations
         vm.startPrank(address(this));
         // Mark Ethereum destination as supported
@@ -174,44 +164,44 @@ contract HashVerificationTest is TestUtils {
         );
         Aori(ARBITRUM_CONTRACT_ADDRESS).addSupportedChain(ETHEREUM_EID);
         vm.stopPrank();
-        
+
         console.log("Test contract: %s", ARBITRUM_CONTRACT_ADDRESS);
         console.log("Test contract EID: %d", ARBITRUM_EID);
         console.log("Signing Hash:           0x%s", toHexString(signingHash));
         console.log("Recovered:       %s", recovered);
-        
+
         // Setup for the deposit
         // Mint tokens to the offerer (signer)
         inputToken.mint(testSigner, order.inputAmount * 2);
-        
+
         // Approve tokens for the deposit
         vm.prank(testSigner);
         inputToken.approve(ARBITRUM_CONTRACT_ADDRESS, order.inputAmount);
-        
+
         // Whitelist solver as an allowed solver
         Aori(ARBITRUM_CONTRACT_ADDRESS).addAllowedSolver(solverAddress);
-        
+
         // Use the original signature that we created for the Arbitrum contract
         console.log("Using signature: 0x%s", toHexString(signature));
-        
+
         // Now attempt the deposit with the solver calling it
         vm.prank(solverAddress);
         Aori(ARBITRUM_CONTRACT_ADDRESS).deposit(order, signature);
-        
+
         // Verify the deposit was successful
         bytes32 orderHash = Aori(ARBITRUM_CONTRACT_ADDRESS).hash(order);
         uint8 status = uint8(Aori(ARBITRUM_CONTRACT_ADDRESS).orderStatus(orderHash));
-        
+
         console.log("\n---- DEPOSIT RESULT ----");
         console.log("Deposit Transaction: SUCCESSFUL");
         console.log("Recovered Signer:    %s", recovered);
         console.log("Transaction Sender:  %s", solverAddress);
         console.log("Order Hash:          0x%s", toHexString(orderHash));
         console.log("Order Status:        %s", status == 1 ? "Active" : "Other");
-        
+
         // Verify the status is Active
         assertEq(status, 1, "Order should be Active after deposit");
-        
+
         console.log("==================================\n");
     }
 
@@ -255,7 +245,9 @@ contract HashVerificationTest is TestUtils {
     /**
      * @notice Helper to convert bytes32 to bytes
      */
-    function bytes32ToBytes(bytes32 data) internal pure returns (bytes memory) {
+    function bytes32ToBytes(
+        bytes32 data
+    ) internal pure returns (bytes memory) {
         bytes memory result = new bytes(32);
         assembly {
             mstore(add(result, 32), data)
@@ -266,17 +258,21 @@ contract HashVerificationTest is TestUtils {
     /**
      * @notice Helper function to convert bytes32 to hex string for logging
      */
-    function toHexString(bytes32 value) internal pure returns (string memory) {
+    function toHexString(
+        bytes32 value
+    ) internal pure returns (string memory) {
         return toHexString(abi.encodePacked(value));
     }
 
     /**
      * @notice Helper function to convert bytes to hex string for logging
      */
-    function toHexString(bytes memory value) internal pure returns (string memory) {
+    function toHexString(
+        bytes memory value
+    ) internal pure returns (string memory) {
         bytes memory alphabet = "0123456789abcdef";
         bytes memory str = new bytes(2 * value.length);
-        for (uint i = 0; i < value.length; i++) {
+        for (uint256 i = 0; i < value.length; i++) {
             str[2 * i] = alphabet[uint8(value[i] >> 4)];
             str[2 * i + 1] = alphabet[uint8(value[i] & 0x0f)];
         }
