+++ test/foundry/25_HookUtilsTests.t.sol
@@ -12,8 +12,8 @@ pragma solidity 0.8.33;
  * 5. test_integration_SrcHook_ValidInvalid - Tests multiple SrcHook scenarios
  * 6. test_integration_DstHook_ValidInvalid - Tests multiple DstHook scenarios
  *
- * This test file verifies that the HookUtils library correctly identifies valid hooks 
- * (those with non-zero addresses) for both SrcHook and DstHook struct types. These utility 
+ * This test file verifies that the HookUtils library correctly identifies valid hooks
+ * (those with non-zero addresses) for both SrcHook and DstHook struct types. These utility
  * functions are critical for validating hook operations within the Aori protocol.
  */
 import "forge-std/Test.sol";
@@ -32,16 +32,20 @@ contract HookTestWrapper {
      * @param hook The SrcHook struct to check
      * @return Whether the hook has a non-zero address
      */
-    function isSomeSrcHook(SrcHook calldata hook) external pure returns (bool) {
+    function isSomeSrcHook(
+        SrcHook calldata hook
+    ) external pure returns (bool) {
         return hook.hookAddress != address(0);
     }
-    
+
     /**
      * @notice Checks if a DstHook is defined (has a non-zero address)
      * @param hook The DstHook struct to check
      * @return Whether the hook has a non-zero address
      */
-    function isSomeDstHook(DstHook calldata hook) external pure returns (bool) {
+    function isSomeDstHook(
+        DstHook calldata hook
+    ) external pure returns (bool) {
         return hook.hookAddress != address(0);
     }
 }
@@ -53,101 +57,85 @@ contract HookTestWrapper {
 contract HookUtilsTest is Test {
     // Test wrapper
     HookTestWrapper public wrapper;
-    
+
     // Test addresses
     address constant ZERO_ADDRESS = address(0);
     address constant TEST_ADDRESS = address(0x1234567890123456789012345678901234567890);
-    
+
     function setUp() public {
         wrapper = new HookTestWrapper();
     }
-    
+
     /**********************************/
     /*     SrcHook Tests             */
     /**********************************/
-    
+
     /// @dev Tests isSome with zero address for SrcHook
     /// @notice Covers line 208 in AoriUtils.sol
     function test_isSome_SrcHook_zeroAddress() public view {
         // Arrange
         SrcHook memory hook = SrcHook({
-            hookAddress: ZERO_ADDRESS,
-            preferredToken: address(0),
-            minPreferedTokenAmountOut: 0,
-            instructions: bytes(""),
-            solver: address(0)
+            hookAddress: ZERO_ADDRESS, preferredToken: address(0), minPreferedTokenAmountOut: 0, instructions: bytes(""), solver: address(0)
         });
-        
+
         // Act
         bool result = wrapper.isSomeSrcHook(hook);
-        
+
         // Assert
         assertFalse(result, "Zero address should return false");
     }
-    
+
     /// @dev Tests isSome with non-zero address for SrcHook
     /// @notice Covers line 208 in AoriUtils.sol
     function test_isSome_SrcHook_nonZeroAddress() public view {
         // Arrange
         SrcHook memory hook = SrcHook({
-            hookAddress: TEST_ADDRESS,
-            preferredToken: address(0),
-            minPreferedTokenAmountOut: 0,
-            instructions: bytes(""),
-            solver: address(0)
+            hookAddress: TEST_ADDRESS, preferredToken: address(0), minPreferedTokenAmountOut: 0, instructions: bytes(""), solver: address(0)
         });
-        
+
         // Act
         bool result = wrapper.isSomeSrcHook(hook);
-        
+
         // Assert
         assertTrue(result, "Non-zero address should return true");
     }
-    
+
     /**********************************/
     /*     DstHook Tests             */
     /**********************************/
-    
+
     /// @dev Tests isSome with zero address for DstHook
     /// @notice Covers line 217 in AoriUtils.sol
     function test_isSome_DstHook_zeroAddress() public view {
         // Arrange
-        DstHook memory hook = DstHook({
-            hookAddress: ZERO_ADDRESS,
-            preferredToken: address(0),
-            preferedDstInputAmount: 0,
-            instructions: bytes("")
-        });
-        
+        DstHook memory hook =
+            DstHook({ hookAddress: ZERO_ADDRESS, preferredToken: address(0), preferedDstInputAmount: 0, instructions: bytes("") });
+
         // Act
         bool result = wrapper.isSomeDstHook(hook);
-        
+
         // Assert
         assertFalse(result, "Zero address should return false");
     }
-    
+
     /// @dev Tests isSome with non-zero address for DstHook
     /// @notice Covers line 217 in AoriUtils.sol
     function test_isSome_DstHook_nonZeroAddress() public view {
         // Arrange
-        DstHook memory hook = DstHook({
-            hookAddress: TEST_ADDRESS,
-            preferredToken: address(0),
-            preferedDstInputAmount: 0,
-            instructions: bytes("")
-        });
-        
+        DstHook memory hook =
+            DstHook({ hookAddress: TEST_ADDRESS, preferredToken: address(0), preferedDstInputAmount: 0, instructions: bytes("") });
+
         // Act
         bool result = wrapper.isSomeDstHook(hook);
-        
+
         // Assert
         assertTrue(result, "Non-zero address should return true");
     }
-    
+
     /**********************************/
     /*     Integration Tests         */
     /**********************************/
-    
+
     /// @dev Tests multiple SrcHook scenarios
     function test_integration_SrcHook_ValidInvalid() public view {
         // Test various hook addresses
@@ -155,13 +143,13 @@ contract HookUtilsTest is Test {
         addresses[0] = ZERO_ADDRESS;
         addresses[1] = TEST_ADDRESS;
         addresses[2] = address(0xDEAD);
-        
+
         bool[] memory expected = new bool[](3);
-        expected[0] = false;  // Zero address
-        expected[1] = true;   // Normal address
-        expected[2] = true;   // Another normal address
-        
-        for (uint i = 0; i < addresses.length; i++) {
+        expected[0] = false; // Zero address
+        expected[1] = true; // Normal address
+        expected[2] = true; // Another normal address
+
+        for (uint256 i = 0; i < addresses.length; i++) {
             SrcHook memory hook = SrcHook({
                 hookAddress: addresses[i],
                 preferredToken: address(0),
@@ -169,12 +157,12 @@ contract HookUtilsTest is Test {
                 instructions: bytes(""),
                 solver: address(0)
             });
-            
+
             bool result = wrapper.isSomeSrcHook(hook);
             assertEq(result, expected[i], string(abi.encodePacked("SrcHook test failed for address: ", addresses[i])));
         }
     }
-    
+
     /// @dev Tests multiple DstHook scenarios
     function test_integration_DstHook_ValidInvalid() public view {
         // Test various hook addresses
@@ -182,20 +170,16 @@ contract HookUtilsTest is Test {
         addresses[0] = ZERO_ADDRESS;
         addresses[1] = TEST_ADDRESS;
         addresses[2] = address(0xDEAD);
-        
+
         bool[] memory expected = new bool[](3);
-        expected[0] = false;  // Zero address
-        expected[1] = true;   // Normal address
-        expected[2] = true;   // Another normal address
-        
-        for (uint i = 0; i < addresses.length; i++) {
-            DstHook memory hook = DstHook({
-                hookAddress: addresses[i],
-                preferredToken: address(0),
-                preferedDstInputAmount: 0,
-                instructions: bytes("")
-            });
-            
+        expected[0] = false; // Zero address
+        expected[1] = true; // Normal address
+        expected[2] = true; // Another normal address
+
+        for (uint256 i = 0; i < addresses.length; i++) {
+            DstHook memory hook =
+                DstHook({ hookAddress: addresses[i], preferredToken: address(0), preferedDstInputAmount: 0, instructions: bytes("") });
+
             bool result = wrapper.isSomeDstHook(hook);
             assertEq(result, expected[i], string(abi.encodePacked("DstHook test failed for address: ", addresses[i])));
         }
